requirements.txt
dnspython
python_dotenv
